+++ public/src/components/App.jsx
@@ -2,6 +2,7 @@ import React, { useState, useEffect } from 'react';
 import Header from './homepage/Header.jsx';
 import Gallery from './homepage/Gallery.jsx';
 import Footer from './homepage/Footer.jsx';
+import Orders from './orders/orders.jsx';
 
 const App = () => {
