+++ public/src/components/orders/borrowedSingle.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+var BorrowedSingle = (props) => {
+  return (
+    <li>{props.data.name} {props.data.author}</li>
+  )
+}
+
+export default BorrowedSingle;
