+++ public/src/components/orders/loanedSingle.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+var LoanedSingle = (props) => {
+  return (
+    <li>{props.data.name} {props.data.author}</li>
+  )
+}
+
+export default LoanedSingle;
