+++ public/src/components/orders/pendingSingle.jsx
@@ -1,0 +1,9 @@
+import React from "react";
+
+var PendingSingle = (props) => {
+  return (
+    <li>{props.data.status} {props.data.name}</li>
+  )
+}
+
+export default PendingSingle;
